should close protocolbuffers#6095

+++ src/google/protobuf/parse_context.h
@@ -125,7 +125,7 @@ class PROTOBUF_EXPORT EpsCopyInputStream {
   // If return value is negative it's an error
   PROTOBUF_MUST_USE_RESULT int PushLimit(const char* ptr, int limit) {
     GOOGLE_DCHECK(limit >= 0);
-    limit += ptr - buffer_end_;
+    limit += static_cast<int>(ptr - buffer_end_);
     limit_end_ = buffer_end_ + (std::min)(0, limit);
     auto old_limit = limit_;
     limit_ = limit;
